Feat: 코스 검색기능 구현 (#99)
* Rename/Fix: 파일 경로 변경, validationUtil 생성

* Feat: 코스 관련 엔티티 생성 (#97)

* Setting: elastic search, kafka 설정 (#97)

* Rename/Feat: 파일 이동, elastic search 를 위한 documented, Repository 생성 (#97)

* Feat: elastic search 를 위한 추가 documented, Repository 생성 (#97)

* Feat: 기능 구현에 필요한 리뷰 enum 생성 (#97)

* Setting: elastic search 추가 설정 (#97)

* Feat: elastic search 에 초기 데이터 구축하도록 구현 (#97)
- publish, consumer 생성
- document 생성, 관련 repository 생성
- 리뷰 생성 기능 구현
- 리뷰 생성 관련 publish, consumer, document, repository 생성

* Feat/Fix: 공통 dto 생성과 일부 dto 구조 변경 (#97)
- 공통 response dto 생성
- reviewCont, isBookmarked 필드 추가
- 이외 화면명세와 맞춰서 필드 수정
** 추후 null 값은 반환하지 않도록 수정해야함
** olleTag 추가해야함 (csv 수정 필요)

* Feat: bookmark 관련 기능 구현 (#97)
- publish, consumer 생성
- 즐겨찾기 생성시 유저 리스트 변경 메서드 작성

* Fix: ollecourse 데이터 수정 완료 (#94)

* Feat: 텍스트 검색 기능  초안 구현 완료 (#97)
- 일부 dto 필드타입 수정
- todo: 더미데이터로 쿼리 확인

* Fix: 텍스트 기반 검색기능 일부 수정 (#97)

* Feat: 위치/태그 검색 구현 (#97)

* Fix: 데이터 중복 추가문제 해결 (#97)

* Feat: 코스/산책로 상세 기능 구현 (#97)

* Fix: 코스/산책로 상세 반환값 일부 수정 (#97)

+++ src/main/java/com/example/jejugudgo/domain/trail/dto/TrailResponseForList.java
@@ -4,6 +4,6 @@ public record TrailResponseForList(
         Long trailId,
         String title,
         double starAvg,
-        Long reviewCount
+        int reviewCount
 ) {
 }
